Don't sync assignment events for gates
Because feature gates aren't needed for A/B test analysis and TestTrack
server will now discard the assignment event in order to allow visitors
to dynamically reassign upon weighting changes, we just
deterministically assign locally each time instead and skip
notification of the event.

This code change is pretty far from the notification code, but leans on
the fact that all notification codepaths rely on
visitor.unsynced_assignments to decide whether to send assignment
notifications, which in turn calls the `unsynced?` method on assignment,
which was previously hard-coded to return true - basically if the rails
client made an assignment, it was considered worthy of syncing to the
server before. Now we will consider feature gate assignments to be
synced, so they will not trigger a notification.

This should make pages and jobs that only have feature gates (and not
experiment assignments) less resource intensive than ever!