Update ActiveRecord::Migration class
This PR updates the migration class name in the migration generator so
it inlcudes the version number if the ActiveRecord version is 5 or
greater. This is needed since you get the following error without it
```StandardError: Directly inheriting from ActiveRecord::Migration is
not supported. Please specify the Rails release the migration was
written for``` When the version is below 5, it will use the standard
class name without a version number.