Keep decided/100% variants at 100% in offline registry (#117)
The offline split registry (used in tests, etc) will now favor variants with weights of 100% when generating deterministic weightings. If no variant has a weight of 100%, then it will fall back on the previous behavior of favoring the variant with the lowest alphanumerically-sorted name (conventionally “false” and “control”).

This means that, when deciding a split via a migration (and/or updating a new variant to 100% in the schema), tests that assumed the previous variant may begin to fail and will require additional code cleanup (or must be explicitly stubbed to use the previous variant).

Because it's technically a breaking change, I've updated the version to `6.0.0`. 

Update: For expediency reasons I also dropped support for Ruby 2.5.