feat: Add Rails 7.1 - 8 support (#141)
* bundle update --bundler

* drop rails 6.1 support

* drop pre ruby 3.2 support

- remove it from the ci matrix
- bump required_ruby_version in the gemspec
- remove call to `ruby` in Gemfile, regen appraisals
- bump TargetRubyVersion in rubocop.yml
  - and rubocop -a, since all of these are HashSyntax changes

* swap rubocop deps for betterlint

* add Rails 7.1 appraisal

- We also rm secrets.yml from the dummy app to deal with a deprecation
warning
- And updated the dummy application.rb to `config.load_defaults` for
Rails 7 and 7.1 (whilst dropping previous versions)

* add Rails 7.2 appraisal

and update `her` code to address a deprecation warning;

https://github.com/rails/rails/blob/v7.2.2.1/activesupport/lib/active_support/proxy_object.rb

* add Rails 8 appraisal

and update the dummy app config, as well as the ci matrix

* have the Rails version range be closed

* bundle update

* bump version.rb

* don't pin concurrent-ruby for 7.1

* re-intro support for ruby 3.1

* exclude ruby 3.1 + Rails 8 combo

Rails 8 requires 3.2

* add missed constraints on railties

* constrain zeitwerk

`railties`'s constraint on zeitwerk is '~> 2.6' (or 2.5 for older
Rails), which means that bundle updating can put us on 2.7, which
requires Ruby 3.2. This will break CI tests for Ruby 3.1.

When we drop support for 3.1, we can unpin this dep (i.e. not even list
it at all, as before).

* make zeitwerk a dev dependency

* use Rails::VERSION::STRING.to_f as arg to config.load_defaults

instead of `rails_version_between`